tech-debt: split demo-data.service.ts orchestrators into helpers (ROK-1142)
Extract batch insert factories into demo-data-batch.utils.ts and the
two install orchestrators (installCoreEntities, installSecondaryEntities)
plus runTasteProfileAggregation into demo-data-install-orchestrate.helpers.ts.

demo-data.service.ts: 333 -> 142 lines (under 300-line cap, target ~150 hit).
DI surface unchanged; behaviour preserved verbatim.
Integration suites pass (admin/demo-data.integration, admin/demo-data-creator-distribution).
